% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_data.R
\name{clear_data}
\alias{clear_data}
\title{Remove data objects from the global environment}
\usage{
clear_data(keep = NULL, gc = FALSE)
}
\arguments{
\item{keep}{Character vector of objects to keep, while everything else will
be removed.}

\item{gc}{Binary. If \code{TRUE}, run garbage collector (\code{gc()}) after clearing
objects.}
}
\value{
Global objects which are not functions are removed from the
environment.
}
\description{
\code{clear_data()} wraps around \code{rm()} to remove data objects from
the global environment, but does not remove sourced functions. It is a
convenient function to put at the end of modular scripts to shed any errant
objects.
}
\details{
It is important to note that this function does not provide a clean
slate restart of R. Full resets should be performed liberally, either by
closing and restarting R or using \code{CTRL/CMD + SHIFT + F10}. Clearing
objects from the environment is does not achieve this. Rather,
\code{clear_data()} is meant to be used at the end of modular scripts. Pertinent
objects can either be saved to .rds or by using the \code{keep} argument.
}
