% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function.R
\name{get_function}
\alias{get_function}
\title{Formatted function file}
\usage{
get_function(
  file_name,
  use_default_dir = FALSE,
  directory = NULL,
  open = TRUE,
  overwrite = FALSE,
  title = TRUE
)
}
\arguments{
\item{file_name}{A name for the function}
}
\value{
Creates a function file as a side effect. Opens the file by default
}
\description{
Creates a formatted script file in the \verb{3_functions} directory by
default.
}
\details{
Just a convenience function to quickly make a function file and open it
}
