% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_markdown.R
\name{get_markdown}
\alias{get_markdown}
\title{Get Markdown}
\usage{
get_markdown(file_name = "markdown", open = TRUE)
}
\arguments{
\item{file_name}{the file name of the markdown to create. Can include the
\code{.Rds} suffix or not.}

\item{open}{if \code{TRUE}, opens the markdown file after creating it.}
}
\value{
a markdown template at path \code{file_name}
}
\description{
Create a template of a markdown file
}
\examples{
get_markdown('my_markdown')
get_markdown('8_markdowns/my_markdown.rds')
}
