% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size.R
\name{get_size}
\alias{get_size}
\title{Get formatted size of an object}
\usage{
get_size(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Arguments passed on to \code{object.size()}}
}
\value{
A formatted character string describing the size of the object.
}
\description{
This function is just a wrapper around \code{object.size()} that checks
how large the object is and uses \code{format()} to provide the output in Kb, Mb, or
Gb as is appropriate.
}
\examples{
get_size(iris)
}
\seealso{
\code{\link[=object.size]{object.size()}}, \code{\link[=format]{format()}}
}
