% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Table with NAs by default}
\usage{
get_table(x, ...)
}
\arguments{
\item{x}{an object that can be interpreted as a factor or list.}

\item{...}{additional arguments passed onto \code{table()}.}
}
\value{
Returns a contingency table.
}
\description{
Wrapper for \code{table()} that with argument \code{useNA = "always"} by
defualt.
}
\details{
This is really better off as an option setting, but I've already
come this far and have gotten used to using it.
}
\seealso{
\code{\link[=table]{table()}}
}
