% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{Print message and current system time}
\usage{
get_time(msg = "Starting at")
}
\arguments{
\item{msg}{a character string, object, or concatenated vector of the two that
will be printed before. They will show before the current system time.}
}
\value{
\code{msg} is concatenated and printed to the console followed by the
current system time.
}
\description{
Print message and current system time
}
\details{
This is a convenient function if you are dealing with long
processing times and loops. It can be placed within a for loop, apply loop,
or map function and take the iteration number to print how far along the
process is.
}
\examples{
get_time()
x <- 426
get_time(c('Starting iteration ', x, ' at'))
}
