% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sandbox.R
\name{remove_sandbox}
\alias{remove_sandbox}
\title{Remove sandbox objects}
\usage{
remove_sandbox()
}
\description{
\code{remove_sandbox} simply removes the objects created by the
\code{build_sandbox} function.
}
\seealso{
\code{\link[=build_sandbox]{build_sandbox()}}
}
