% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folders.R
\name{get_folders}
\alias{get_folders}
\title{get_folders}
\usage{
get_folders(folder_names = NULL)
}
\arguments{
\item{folder_names}{is a character vector names for folders.}
}
\description{
Create folders in project root
}
