% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_readme.R
\name{get_readme}
\alias{get_readme}
\title{Get README}
\usage{
get_readme()
}
\value{
creates 'README.Rmd' in root folder
}
\description{
Create a template of a README file with some options that are objectively
convenient.
}
\examples{
get_readme()
}
