% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script.R
\name{get_script}
\alias{get_script}
\title{Formatted script file}
\usage{
get_script(
  file_name,
  directory = "4_scripts/",
  open = TRUE,
  overwrite = FALSE,
  title = TRUE
)
}
\arguments{
\item{file_name}{A file name for}

\item{...}{additional arguments passed onto \code{table()}.}
}
\value{
Creates a script file as a side effect. Opens the script by default
}
\description{
Creates a formatted script file in the \verb{4_scripts} directory by
default.
}
\details{
Just a convenience function to quickly make a script file and open it
}
