% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_str.R
\name{get_str}
\alias{get_str}
\title{User friendly wrapper for \code{str()}}
\usage{
get_str(obj, level = 2, all_cols = TRUE, strict.width = "cut", ...)
}
\arguments{
\item{obj}{an R object.}

\item{level}{if \code{obj} is a nested list, how many layers down should be
shown.}

\item{all_cols}{is passed on to the \code{max.levels} argument of \code{str()}. When
\code{obj} is a data frame with many columns, TRUE means that all columns
should be shown in the output. FALSE means that}

\item{strict.width}{is passed on to the \code{strict.width} argument of \code{str()}.
The default, "cut", means that the output will be cut at the edge of the
current size of the console, no matter how many or few values it will show.}

\item{...}{additional arguments passed on to \code{str()}.}
}
\value{
Returns the usual output of \code{str()}, but formatted to be clearer,
show more information at once, and avoid typing in the rather verbose
arguments into \code{str()},
}
\description{
User friendly wrapper for \code{str()}
}
\examples{
get_str(iris)
get_str(mtcars)
}
